% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{movies_new}
\alias{movies_new}
\title{New movies data}
\usage{
data(movies_new)
}
\value{
A tibble
}
\description{
Some data for the workshop
}
\examples{
data(movies_new)
}
\keyword{datasets}
